+++ src/convert.jl
@@ -60,16 +60,6 @@ and not to a `Vector`. There should not be more than one canonical conversion ru
 given Python type.
 
 Other priorities are reserved for internal use.
-
-### Optional optimization (for experts)
-
-The `func` is given the only reference to `x`. This means it may call `pydel!(x)` when done
-with `x` provided `x` is not referenced anywhere else, such as in the return value. See
-[`pydel!`](@ref).
-
-If you are returning a wrapper type, such as `PyList(x)`, then the object `x` should be
-duplicated first, as in `Py(x)`. It is recommended to force this in the inner constructor
-of the wrapper type.
 """
 function pyconvert_add_rule(pytypename::String, type::Type, func::Function, priority::PyConvertPriority=PYCONVERT_PRIORITY_NORMAL)
     @nospecialize type func
@@ -305,29 +295,29 @@ function pyconvert_rule_fast(::Type{T}, x::Py) where {T}
         pyisFalse(x) && return pyconvert_return(false)
         pyisTrue(x) && return pyconvert_return(true)
     elseif (T == Int) | (T == BigInt)
-        pyisint(x) && return pyconvert_rule_int(T, Py(x))
+        pyisint(x) && return pyconvert_rule_int(T, x)
     elseif (T == Float64)
         pyisfloat(x) && return pyconvert_return(T(pyfloat_asdouble(x)))
     elseif (T == ComplexF64)
         pyiscomplex(x) && return pyconvert_return(T(pycomplex_ascomplex(x)))
     elseif (T == String) | (T == Char) | (T == Symbol)
-        pyisstr(x) && return pyconvert_rule_str(T, Py(x))
+        pyisstr(x) && return pyconvert_rule_str(T, x)
     elseif (T == Vector{UInt8}) | (T == Base.CodeUnits{UInt8,String})
-        pyisbytes(x) && return pyconvert_rule_bytes(T, Py(x))
+        pyisbytes(x) && return pyconvert_rule_bytes(T, x)
     elseif (T <: StepRange) | (T <: UnitRange)
-        pyisrange(x) && return pyconvert_rule_range(T, Py(x))
+        pyisrange(x) && return pyconvert_rule_range(T, x)
     end
     pyconvert_unconverted()
 end
 
 function pytryconvert(::Type{T}, x_) where {T}
-    # Copy the input
+    # Convert the input to a Py
     x = Py(x_)
 
     # We can optimize the conversion for some types by overloading pytryconvert_fast.
     # It MUST give the same results as via the slower route using rules.
     ans1 = pyconvert_rule_fast(T, x) :: pyconvert_returntype(T)
-    pyconvert_isunconverted(ans1) || (pydel!(x); return ans1)
+    pyconvert_isunconverted(ans1) || return ans1
 
     # get rules from the cache
     # TODO: we should hold weak references and clear the cache if types get deleted
@@ -342,12 +332,10 @@ function pytryconvert(::Type{T}, x_) where {T}
 
     # apply the rules
     for rule in rules
-        ans2 = rule(Py(x)) :: pyconvert_returntype(T)
-        pyconvert_isunconverted(ans2) || (pydel!(x); return ans2)
+        ans2 = rule(x) :: pyconvert_returntype(T)
+        pyconvert_isunconverted(ans2) || return ans2
     end
 
-    # if no rule succeeded, assume nothing references x
-    pydel!(x)
     return pyconvert_unconverted()
 end
 
@@ -395,24 +383,6 @@ pyconvertarg(::Type{T}, x, name) where {T} = @autopy x @pyconvert T x_ begin
     pythrow()
 end
 
-macro pyconvert_and_del(T, x, onfail=:(return $pyconvert_unconverted()))
-    quote
-        ans = pytryconvert($(esc(T)), $(esc(x)))
-        pydel!($(esc(x)))
-        if pyconvert_isunconverted(ans)
-            $(esc(onfail))
-        else
-            pyconvert_result($(esc(T)), ans)
-        end
-    end
-end
-
-pyconvert_and_del(::Type{T}, x) where {T} = begin
-    ans = pyconvert(T, x)
-    pydel!(x)
-    ans
-end
-
 function init_pyconvert()
     push!(PYCONVERT_EXTRATYPES, pyimport("io"=>"IOBase"))
     push!(PYCONVERT_EXTRATYPES, pyimport("numbers"=>("Number", "Complex", "Real", "Rational", "Integral"))...)
@@ -444,6 +414,7 @@ function init_pyconvert()
     pyconvert_add_rule("io:IOBase", PyIO, pyconvert_rule_io, priority)
     pyconvert_add_rule("_io:_IOBase", PyIO, pyconvert_rule_io, priority)
     pyconvert_add_rule("pandas.core.frame:DataFrame", PyPandasDataFrame, pyconvert_rule_pandasdataframe, priority)
+    pyconvert_add_rule("pandas.core.arrays.base:ExtensionArray", PyList, pyconvert_rule_sequence, priority)
     pyconvert_add_rule("builtins:tuple", Tuple, pyconvert_rule_iterable, priority)
     pyconvert_add_rule("datetime:datetime", DateTime, pyconvert_rule_datetime, priority)
     pyconvert_add_rule("datetime:date", Date, pyconvert_rule_date, priority)